+++ src/main/java/org/avengers/capstone/hostelrenting/repository/RenterRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RenterRepository extends JpaRepository<Renter, Integer> {
+public interface RenterRepository extends JpaRepository<Renter, Long> {
     Optional<Renter> findByPhone(String phone);
 }
